% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{A function to calculate contrasts}
\usage{
contrast(linearmodel, K)
}
\arguments{
\item{linearmodel}{A linear model.}
}
\description{
A function to calculate contrasts
}
\examples{
contrast(lm(dv ~ iv1*iv2, data = data), c(0,0,0,1))
}
\keyword{contrast}
