% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type3anova.R
\name{type3anova}
\alias{type3anova}
\title{A function for Type III Anova (generates the same output as SPSS does)}
\usage{
type3anova(linearmodel)
}
\arguments{
\item{linearmodel}{A linear model.}
}
\description{
A function for Type III Anova (generates the same output as SPSS does)
}
\examples{
type3anova(lm(y ~ x1 * x2, data=dataset))
}
\keyword{anova}
\keyword{type3}
